refactor(commands): split commands/profile.rs into per-domain submodules
Convert 1663-line monolith to commands/profile/ directory with shared.rs,
lifecycle.rs, optimizations.rs, favorites.rs, config_history.rs, and a
tests/ subdirectory. All 22 Tauri command snake_case names preserved.
capture_config_revision re-exported for commands/protondb.rs.

Closes #309
Refs #290 (task 8)
