+++ src/crosshook-native/src-tauri/src/commands/profile/mod.rs
@@ -1,0 +1,50 @@
+//! Profile Tauri commands, split into per-domain submodules.
+//!
+//! - `shared`         — helpers, DTOs, and constants shared across submodules
+//! - `lifecycle`      — CRUD and import/export commands
+//! - `optimizations`  — launch optimization and preset commands
+//! - `favorites`      — favorite profile commands
+//! - `config_history` — config revision history, diff, rollback, and mark-known-good
+
+mod config_history;
+mod favorites;
+mod lifecycle;
+mod optimizations;
+mod shared;
+
+#[cfg(test)]
+mod tests;
+
+pub use config_history::{
+    profile_config_diff, profile_config_history, profile_config_rollback, profile_mark_known_good,
+};
+pub use favorites::{profile_list_favorites, profile_set_favorite};
+pub use lifecycle::{
+    profile_delete, profile_duplicate, profile_export_toml, profile_import_legacy, profile_list,
+    profile_list_summaries, profile_load, profile_rename, profile_save,
+};
+pub use optimizations::{
+    profile_apply_bundled_optimization_preset, profile_list_bundled_optimization_presets,
+    profile_save_gamescope_config, profile_save_launch_optimizations, profile_save_mangohud_config,
+    profile_save_manual_optimization_preset, profile_save_trainer_gamescope_config,
+};
+pub use shared::capture_config_revision;
+
+// Re-export the `__cmd__` companion items generated by `#[tauri::command]` so
+// that `generate_handler!` can resolve them via `commands::profile::<name>`.
+pub use config_history::{
+    __cmd__profile_config_diff, __cmd__profile_config_history, __cmd__profile_config_rollback,
+    __cmd__profile_mark_known_good,
+};
+pub use favorites::{__cmd__profile_list_favorites, __cmd__profile_set_favorite};
+pub use lifecycle::{
+    __cmd__profile_delete, __cmd__profile_duplicate, __cmd__profile_export_toml,
+    __cmd__profile_import_legacy, __cmd__profile_list, __cmd__profile_list_summaries,
+    __cmd__profile_load, __cmd__profile_rename, __cmd__profile_save,
+};
+pub use optimizations::{
+    __cmd__profile_apply_bundled_optimization_preset,
+    __cmd__profile_list_bundled_optimization_presets, __cmd__profile_save_gamescope_config,
+    __cmd__profile_save_launch_optimizations, __cmd__profile_save_mangohud_config,
+    __cmd__profile_save_manual_optimization_preset, __cmd__profile_save_trainer_gamescope_config,
+};
