+++ src/crosshook-native/src-tauri/src/commands/profile/tests/lifecycle.rs
@@ -1,0 +1,109 @@
+use crosshook_core::export::check_launcher_exists;
+use crosshook_core::profile::{
+    GameProfile, GameSection, LaunchSection, LauncherSection, SteamSection, TrainerLoadingMode,
+    TrainerSection,
+};
+use std::fs;
+use tempfile::tempdir;
+
+use crate::commands::profile::shared::cleanup_launchers_for_profile_delete;
+
+fn steam_profile(home: &str) -> GameProfile {
+    GameProfile {
+        game: GameSection {
+            name: "Test Game".to_string(),
+            executable_path: String::new(),
+            custom_cover_art_path: String::new(),
+            custom_portrait_art_path: String::new(),
+            custom_background_art_path: String::new(),
+        },
+        trainer: TrainerSection {
+            path: "/tmp/trainers/test.exe".to_string(),
+            kind: String::new(),
+            loading_mode: TrainerLoadingMode::SourceDirectory,
+            trainer_type: "unknown".to_string(),
+            required_protontricks: Vec::new(),
+            community_trainer_sha256: String::new(),
+        },
+        steam: SteamSection {
+            app_id: "12345".to_string(),
+            compatdata_path: format!("{home}/.local/share/Steam/steamapps/compatdata/12345"),
+            launcher: LauncherSection {
+                display_name: "Test Game".to_string(),
+                icon_path: String::new(),
+            },
+            ..Default::default()
+        },
+        launch: LaunchSection {
+            method: "steam_applaunch".to_string(),
+            ..Default::default()
+        },
+        ..Default::default()
+    }
+}
+
+fn create_watermarked_launcher_files(script_path: &str, desktop_path: &str) {
+    fs::create_dir_all(
+        std::path::Path::new(script_path)
+            .parent()
+            .expect("script parent"),
+    )
+    .expect("script dirs");
+    fs::create_dir_all(
+        std::path::Path::new(desktop_path)
+            .parent()
+            .expect("desktop parent"),
+    )
+    .expect("desktop dirs");
+    fs::write(
+        script_path,
+        "#!/usr/bin/env bash\n# Generated by CrossHook\n",
+    )
+    .expect("write script");
+    fs::write(
+        desktop_path,
+        "[Desktop Entry]\nName=Test Game - Trainer\nComment=Generated by CrossHook\n",
+    )
+    .expect("write desktop");
+}
+
+#[test]
+fn cleanup_launchers_for_profile_delete_uses_derived_steam_paths() {
+    let temp = tempdir().expect("temp dir");
+    let home = temp.path().to_string_lossy().into_owned();
+    let profile = steam_profile(&home);
+    let steam_root = format!("{home}/.local/share/Steam");
+
+    let info = check_launcher_exists(
+        &profile.steam.launcher.display_name,
+        &profile.steam.app_id,
+        &profile.trainer.path,
+        &home,
+        &steam_root,
+    )
+    .expect("check launcher exists");
+    create_watermarked_launcher_files(&info.script_path, &info.desktop_entry_path);
+
+    let result = cleanup_launchers_for_profile_delete("test-profile", &profile)
+        .expect("cleanup should succeed");
+
+    assert!(result.is_some());
+    assert!(!std::path::Path::new(&info.script_path).exists());
+    assert!(!std::path::Path::new(&info.desktop_entry_path).exists());
+}
+
+#[test]
+fn cleanup_launchers_for_profile_delete_skips_native_profiles() {
+    let profile = GameProfile {
+        launch: LaunchSection {
+            method: "native".to_string(),
+            ..Default::default()
+        },
+        ..GameProfile::default()
+    };
+
+    let result = cleanup_launchers_for_profile_delete("native-profile", &profile)
+        .expect("native cleanup should not fail");
+
+    assert!(result.is_none());
+}
