+++ src/crosshook-native/src-tauri/src/commands/profile/tests/mod.rs
@@ -1,0 +1,2 @@
+mod lifecycle;
+mod optimizations;
